We were facing connection fd leak when connection goes in destroyed state, which causing clients hung, as fd was not rearmed for epoll, client will get tcp zero window and may or mat not do connection reset based on kernel version.
If connection xprt is in destroyed state, then we may not
be able process POLLOUT even for send task and refs taken will
not go away.
We don't need to register for POLLOUT event with ref taken for send
queue, as we take ref for send task.
We don't need ref for each response we queue, we can just single ref
for send queue and release it when we start processing.
We need to close xp_fd_send on connection destroy.